Feature: Ermaessigung als Gleitkommazahl (Erst- und Zweitkraft)
Im Bearbeitungsmodus lassen sich Stundenermaessigungen jetzt in halben
Stunden setzen (+/- in 0,5-Schritten, beliebige Dezimaleingabe moeglich).

DB: ermaessigung UND upz werden in erstkraft/zweitkraft von INT(11) auf
DECIMAL(6,2) umgestellt - neu ueber 01-schema.sql, bestehende Datenbanken
ueber die Selbstheilung elli_ensure_dezimal_stunden() (ALTER nur, wenn die
Spalte noch kein DECIMAL ist). upz muss mit, weil es sich aus
Regelstundenmass minus Ermaessigung ergibt und sonst gerundet wuerde.

API: Ermaessigung/UPZ werden als Zahl statt als DECIMAL-String geliefert;
in den Word-Vorlagen laufen sie durch $fmtStunden (2,5 statt 2.50). Der
Helfer schnitt bisher nur eine Null ab (2,00 -> 2,0), jetzt korrigiert.

Frontend: fehlende Methode validateErmaessigung ergaenzt (war verdrahtet,
aber nie definiert), stepErmaessigung mit Rundung auf zwei Nachkommastellen,
Zweitkraft-Feld auf v-model.number, Speichern normalisiert die Werte.

+++ docker/db/init/01-schema.sql
@@ -46,8 +46,9 @@ CREATE TABLE IF NOT EXISTS `erstkraft` (
   `kuerzel`            VARCHAR(10)  DEFAULT NULL,
   `farbe`              VARCHAR(7)   DEFAULT NULL,
   `pflichtstunden`     INT(11)      DEFAULT 0,
-  `ermaessigung`       INT(11)      DEFAULT 0,
-  `upz`                INT(11)      DEFAULT 0,
+  -- Ermaessigung und UPZ als Dezimalwerte (halbe/viertel Stunden moeglich)
+  `ermaessigung`       DECIMAL(6,2) DEFAULT 0,
+  `upz`                DECIMAL(6,2) DEFAULT 0,
   `faecher`            VARCHAR(255) DEFAULT NULL,
   `textfarbe`          VARCHAR(7)   DEFAULT '#ffffff',
   `ermaessigung_grund` VARCHAR(255) DEFAULT NULL,
@@ -66,9 +67,10 @@ CREATE TABLE IF NOT EXISTS `zweitkraft` (
   `kuerzel`            VARCHAR(10)  DEFAULT NULL,
   `farbe`              VARCHAR(7)   DEFAULT NULL,
   `textfarbe`          VARCHAR(7)   DEFAULT '#ffffff',
-  `ermaessigung`       INT(11)      DEFAULT 0,
+  -- Ermaessigung und UPZ als Dezimalwerte (halbe/viertel Stunden moeglich)
+  `ermaessigung`       DECIMAL(6,2) DEFAULT 0,
   `grund_ermaessigung` VARCHAR(255) DEFAULT NULL,
-  `upz`                INT(11)      DEFAULT 0,
+  `upz`                DECIMAL(6,2) DEFAULT 0,
   PRIMARY KEY (`id`),
   KEY `idx_zweitkraft_schuljahr` (`schuljahr_id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
